+++ odoo/local-src/specific_account/models/__init__.py
@@ -11,3 +11,4 @@ from . import sale_order
 from . import product_product
 from . import account_move_line
 from . import res_partner
+from . import stock_picking